Version 1.1
Although functionally the same as version 1.0 the code has been overhauled to improve the ease of adding future features. Both massive and massless bodies are now implemented as objects, storing information relevant to the particular body. This will allow easy creation of multiple bodies in future versions including the possibility of modelling the dynamics of massive bodies. The coordinate system is no longer based on the centre of the Earth so this also permits the future modelling of the Moon orbiting the Earth. An error with the Runge-Kutta method was also fixed; the multipliers of the K vector components on lines 36 - 39 were incorrect which has now been addressed.